+++ lib/state_lifecycle_observer.dart
@@ -1,4 +1,4 @@
-export 'src/observer_mixin.dart';
+export 'src/owner_mixin.dart';
 export 'src/lifecycle_observer.dart';
 
 export 'src/observer/anim.dart';